requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
httpx
python-dotenv
